requirements.txt
pyelftools==0.29
pyserial==3.5
bleak==0.22.3